+++ projects/chatbot-ui/src/client/App.tsx
@@ -2,13 +2,11 @@ import { Header } from '#/components/Header'
 import { MessageArea } from '#/components/MessageArea'
 import { MessageAreaScroll } from '#/components/MessageAreaScroll'
 import { MessageInput } from '#/components/MessageInput'
+import { useChat } from '#/hooks/useChat'
 import { useTheme } from '#/hooks/useTheme'
-import { useChat } from './useChat'
 
 export function App() {
-  // テーマの初期化
   useTheme()
-
   const chat = useChat()
   return (
     <div className='flex h-dvh flex-col bg-background'>
